Reify and cleanup v1 wire protocol
